requirements.txt
beautifulsoup4
selenium
requests
mysql-connector-python
currencyconverter
pandas
python-dotenv